Element: Change Path definition
Redefining what a Path is. This is to help remove
confusion and overlap between Path, BoundedPath and
PathFunc definitions. There is now just a PathFunc for
defining paths, and a Path namespace for applying
various common transformations to PathFuncs.

There is a StandardSensions library containing commonly
used path definitions.

A lot of code in this commit is commented out, and will
be re-added as unit tests are written to ensure the outputs
are correct.