Element: Remove memberwise from Prelude
We decided that memberwise is not a good idea. Initially
it was conceived as a way to make Vector operators easier
to write, but it's not much benefit in Vector2-4. For
VectorN, any dimensional vectors, they would be implemented
using lists to store the elements. In that case, iterating
over lists and using List.zip is sufficient.