Replace defstack checks with a simpler mechanism
Previously the defstack was used to ensure we weren't trying to validate
the same def recursively, e.g.:
  validate local def
  -> validate code
  -> extcall to itself
  -> validate local def ...

This was used because the EXTCALL validation must verify that the
def being pointed to is valid before continuing, but we don't yet
know whether the def is extern (that was being checked later in the
EXTCALL validation logic). Instead, we now pass a set of required flags
into the validate_def function, allowing us to early-exit the validation
if the def *isn't* an extern function, removing the possibility of an
infinite loop.